+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Removed ubiquitous subscript indexing, now there's a dedicated `.` operator for rearranging the stack.
+
 ## [0.1.0] - 2026-07-05
 Initial release